fix bug

+++ daemon/src/daemon.c
@@ -273,7 +273,7 @@ void RefreshEvents( event_fifo_t * events )
     static struct timespec last_nano_tick;
 
     timespec_get( &current_nano_tick, TIME_UTC );
-    if( ( current_nano_tick.tv_nsec - last_nano_tick.tv_nsec ) > 500000000L )
+    if( (unsigned long)( current_nano_tick.tv_nsec - last_nano_tick.tv_nsec ) > 500000000UL )
     {
         last_nano_tick = current_nano_tick;
         FIFO_Enqueue( events, EVENT( Heartbeat ) );
